styling match general design + add 2 article for each mood
